fix: Moved components, per #409
