+++ src/components/pages/data/state-nav.js
@@ -1,16 +1,94 @@
-import React from 'react'
-import { Link } from 'gatsby'
-import { Box } from '../../common/flexbox'
+/* eslint jsx-a11y/label-has-associated-control: 0 */
 
-export default ({ stateList }) => (
-  <Box width={[1, 1, 1, 1 / 2]} className="state-nav" m="0 auto 1rem">
-    <h3>Jump to a state:</h3>
-    <ul>
-      {stateList.map(state => (
-        <li key={state.node.state}>
-          <Link to={`/data#state-${state.node.state}`}>{state.node.state}</Link>
-        </li>
-      ))}
-    </ul>
-  </Box>
-)
+import React, { useState } from 'react'
+import {
+  Combobox,
+  ComboboxInput,
+  ComboboxPopover,
+  ComboboxList,
+  ComboboxOption,
+} from '@reach/combobox'
+import '@reach/combobox/styles.css'
+
+export default ({ stateList }) => {
+  const [searchTerm, setSearchTerm] = useState('')
+
+  const searchStates = term => {
+    if (term.trim() === '') {
+      return null
+    }
+    const results = []
+    stateList.forEach(({ node }) => {
+      if (node.name.toLowerCase().search(term.toLowerCase().trim()) === 0) {
+        results.push(node)
+      }
+    })
+    return results
+  }
+
+  const results = searchStates(searchTerm)
+
+  return (
+    <div className="state-combobox-nav">
+      <noscript>
+        <style>{`
+    .state-combobox-nav {
+      display: none !important;
+    }
+    `}</style>
+      </noscript>
+      <label htmlFor="jump-to-state">
+        Type a state&apos;s name to jump to it
+      </label>
+      <Combobox
+        openOnFocus
+        onSelect={selectedItem => {
+          const stateId = stateList.find(({ node }) => {
+            return node.name === selectedItem
+          })
+          if (stateId && typeof window !== 'undefined') {
+            window.location.hash = `state-${stateId.node.state.toLowerCase()}`
+          }
+        }}
+      >
+        <ComboboxInput
+          id="jump-to-state"
+          placeholder="State or territory"
+          autoComplete="off"
+          onChange={event => {
+            setSearchTerm(event.target.value)
+          }}
+        />
+        {results ? (
+          <ComboboxPopover className="state-combobox-popover">
+            {results.length > 0 ? (
+              <ComboboxList aria-label="States">
+                {results.slice(0, 10).map(result => (
+                  <ComboboxOption
+                    key={`state-search-${result.state}`}
+                    value={result.name}
+                  />
+                ))}
+              </ComboboxList>
+            ) : (
+              <span style={{ display: 'block', margin: 8 }}>
+                No states found
+              </span>
+            )}
+          </ComboboxPopover>
+        ) : (
+          <ComboboxPopover className="state-combobox-popover">
+            <ComboboxList>
+              {stateList.map(({ node }) => (
+                <ComboboxOption
+                  key={`state-search-${node.state}`}
+                  value={node.name}
+                />
+              ))}
+            </ComboboxList>
+          </ComboboxPopover>
+        )}
+      </Combobox>
+    </div>
+  )
+}
